Refactor scaled_mm into separate file and consolidation of utility functions for CPU/CUDA/XPU
