+++ aten/src/ATen/native/mkldnn/xpu/ScaledBlas.cpp
@@ -5,7 +5,7 @@
 #include <ATen/native/Resize.h>
 #include <ATen/native/mkldnn/xpu/detail/oneDNN.h>
 #include <ATen/native/xpu/Blas.h>
-#include <ATen/xpu/XPUScaledBlas.h>
+#include <ATen/native/ScaledBlasUtils.h>
 #include <torch/library.h>
 
 #ifndef AT_PER_OPERATOR_HEADERS
@@ -359,7 +359,7 @@ using acceptance_fn = std::function<bool(
     ArrayRef<Tensor>&)>;
 using namespace std::placeholders;
 
-namespace scaled_blas = at::native::onednn::scaled;
+namespace scaled_blas = at::scaled;
 using scaled_blas::convert_int_to_enum;
 using scaled_blas::ScaledGemmImplementation;
